allow use with CGI::Simple
If the passed in obj is CGI::Simple rather than CGI, the
$q->tmpFileName() method will not be available. Instead Simple
provides a filehandle which we can use to get the file
info from, using File::MMagic.

Initially I had tried to use 
$q->can("tmpFileName") 
to check if the method was available, but it turns out that that
will always return undef. So instead I've used
$q->isa("CGI::Simple") 
and then fetch the filehandle if that's true, otherwise assuming,
as before, that it's a CGI object.